requirements.txt
attrs==20.3.0
cffi==1.14.4
click==7.1.2
cryptography==3.2.1
Flask==1.1.2
flask-jsonschema-validator==0.0.4
itsdangerous==1.1.0
Jinja2==2.11.2
jsonschema==3.2.0
MarkupSafe==1.1.1
pycparser==2.20
PyJWT==1.7.1
pymongo==3.11.1
pyrsistent==0.17.3
six==1.15.0
Werkzeug==1.0.1
